create separate component for this to make the code easier to move
